pyproject.toml
# Copyright 2025 Canonical Ltd.
# See LICENSE file for licensing details.

[project]
name = "penpot-operator"
version = "0.0.0"
description = "An operator deploying and managing Penpot."
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
]
dependencies = [
  "cosl==1.4.0",
  "dnspython==2.8.0",
  "jsonschema==4.26.0",
  "ops==3.5.2",
  "pydantic==2.12.5",
  "requests==2.32.5",
]

[dependency-groups]
fmt = [
  "ruff",
]
lint = [
  "boto3",
  "codespell",
  "jubilant~=1.0",
  "mypy",
  "oauth-tools @ git+https://github.com/canonical/iam-bundle@6d1fc10e6a489d693e33a6ebc89d3c80c1c35aa9#egg=oauth_tools",
  "ops-scenario",
  "pep8-naming",
  "pytest",
  "pytest-asyncio",
  "pytest-operator",
  "pytest-playwright",
  "ruff",
  "types-pyyaml",
  "types-requests",
]
unit = [
  "coverage[toml]",
  "ops-scenario",
  "pytest",
]
coverage-report = [
  "coverage[toml]",
  "pytest",
]
static = [
  "bandit[toml]",
]
integration = [
  "allure-pytest>=2.8.18",
  "allure-pytest-collection-report @ git+https://github.com/canonical/data-platform-workflows@v24.0.0#subdirectory=python/pytest_plugins/allure_pytest_collection_report",
  "boto3",
  "jubilant~=1.0",
  "juju==3.6.1.3",
  "oauth-tools @ git+https://github.com/canonical/iam-bundle@6d1fc10e6a489d693e33a6ebc89d3c80c1c35aa9#egg=oauth_tools",
  "pytest",
  "pytest-asyncio",
  "pytest-operator",
  "pytest-playwright",
]

[tool.uv]
package = false

[tool.ruff]
target-version = "py310"
line-length = 99

# enable ruff linters:
#   A flake8-builtins
#   B flake8-bugbear
# CPY flake8-copyright
#   D pydocstyle
# E/W pycodestyle
#   F Pyflakes
#   I isort
#   N pep8-naming
# RUF Ruff-specific rules
#   S flake8-bandit
# SIM flake8-simplify
#  TC flake8-type-checking
#  UP pyupgrade
lint.select = [ "A", "B", "C", "CPY", "D", "E", "F", "I", "N", "RUF", "S", "SIM", "TC", "UP", "W" ]
lint.ignore = [
  "B904",
  "D107",
  "D203",
  "D204",
  "D205",
  "D213",
  "D215",
  "D400",
  "D404",
  "D406",
  "D407",
  "D408",
  "D409",
  "D413",
  "E501",
  "S105",
  "S603",
  "TC002",
  "TC006",
  "UP006",
  "UP007",
  "UP035",
  "UP045",
]
lint.per-file-ignores."tests/*" = [ "B011", "D100", "D101", "D102", "D103", "D104", "D212", "D415", "D417", "S" ]
lint.flake8-copyright.author = "Canonical Ltd."
lint.flake8-copyright.min-file-size = 1
lint.flake8-copyright.notice-rgx = "Copyright\\s\\d{4}([-,]\\d{4})*\\s+"
lint.mccabe.max-complexity = 10
lint.pydocstyle.convention = "google"

[tool.codespell]
skip = "build,lib,venv,icon.svg,.tox,.git,.mypy_cache,.ruff_cache,.coverage,htmlcov,uv.lock,grafana_dashboards"
ignore-words-list = "edn"

[tool.pytest.ini_options]
minversion = "6.0"
log_cli_level = "INFO"

[tool.coverage.run]
branch = true

[tool.coverage.report]
show_missing = true

[tool.mypy]
check_untyped_defs = true
disallow_untyped_defs = true
explicit_package_bases = true
ignore_missing_imports = true
namespace_packages = true

[[tool.mypy.overrides]]
disallow_untyped_defs = false
module = "tests.*"

[tool.bandit]
exclude_dirs = [ "/venv/" ]

[tool.bandit.assert_used]
skips = [ "*/*test.py", "*/test_*.py", "*tests/*.py" ]
